Undo rename to rcdt_utilities_py.
