+++ rcdt_utilities/CMakeLists.txt
@@ -42,7 +42,7 @@ install(PROGRAMS
 )
 
 # Install python package
-ament_python_install_package(${PROJECT_NAME}_py)
+ament_python_install_package(${PROJECT_NAME})
 
 if(BUILD_TESTING)
   find_package(ament_lint_auto REQUIRED)
