+++ rcdt_utilities/package.xml
@@ -20,6 +20,8 @@ SPDX-License-Identifier: Apache-2.0
   <buildtool_depend>rosidl_default_generators</buildtool_depend>
   <member_of_group>rosidl_interface_packages</member_of_group>
 
+  <depend>rcdt_utilities_msgs</depend>
+
   <test_depend>ament_lint_auto</test_depend>
 
   <export>
